another try to get multiproc stuff to work on windows
